pyproject.toml
[tool.poetry]
name = "order-management-service"
version = "0.1.0"
description = ""
authors = ["Anton Gromov <seed14112@yaandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^6.0.0"
black = "^25.1.0"
pre-commit = "^4.1.0"
coverage = "^7.6.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = """
    migrations/
"""

[tool.isort]
include_trailing_comma = true
line_length = 119
lines_after_imports = 2
multi_line_output = 3
skip = ["__init__.py"]
